requirements.txt
charset-normalizer
numpy
opencv-python
portalocker
pyyaml
windows-capture
